apply suggestions from review
