+++ react-instantsearch-hooks/facet-dropdown/src/hooks/useCloseDropdown.ts
