+++ react-instantsearch-hooks/facet-dropdown/src/hooks/useLockedBody.ts
