+++ react-instantsearch-hooks/facet-dropdown/src/hooks/useMediaQuery.ts
